Cardioish: A Symbolic Language of Cardiology
**Objective:**
The primary aim of this project is to enable a deeper and more explainable analysis of ECG signals to contribute to the early detection, accurate diagnosis, and monitoring of heart diseases. To achieve this goal, the Cardioish language and the XFE (Explainable Feature Extraction) model are used together. Cardioish represents ECG leads with symbols that provide insights into different regions of the heart. These symbols make the region-specific electrical activities, which are associated with specific cardiac diseases and conditions, more comprehensible.

**Early Detection and Diagnosis of Diseases:**
The Cardioish language and XFE model aim to detect certain cardiac disorders early by identifying characteristic changes in ECG signals. For example, the diagnosis of conditions such as myocardial infarction (heart attack) or atrial fibrillation (AF) can be conducted quickly and regionally using this language. Early detection is critical, especially for preventing the progression of acute and chronic conditions.

**Ensuring Explainability and Interpretability:**
The XFE model used in this project extracts distinctive features from ECG signals in an explainable manner. The symbols within the Cardioish language represent which part of the ECG signal corresponds to which cardiac region and indicate functional abnormalities in that region. This explainability allows physicians and specialists to understand not only the results but also the underlying reasons behind those results.

**Understanding Complex Diseases through Causal Cardiac Connectivity (CCCT):**
The Causal Cardiac Connectivity Theory (CCCT) aims to analyze causal relationships and connections between different regions of the heart. How does an electrical change in one specific region of the heart affect other regions? CCCT aids in analyzing these connections and transitions, providing insights into how disease spreads within the heart and which regions it disrupts. This approach is particularly valuable for understanding the effects of complex diseases, such as myocarditis (inflammation of the heart muscle), where the Cardioish language offers a more detailed map alongside CCCT.

**Conclusion:**
This project enhances understanding in the field of cardiac health by enabling us to glean more information from ECG data. In doing so, it contributes to the earlier diagnosis of heart diseases, supports the development of personalized treatments, and strengthens clinical decision-support systems.